+++ packages/braintree-integration/src/braintree-paypal/braintree-paypal-customer-strategy.spec.ts
@@ -25,11 +25,16 @@ import {
     PaymentIntegrationService,
     PaymentMethod,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
-import { PaymentIntegrationServiceMock } from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
+import {
+    getConfig,
+    PaymentIntegrationServiceMock,
+} from '@bigcommerce/checkout-sdk/payment-integrations-test-utils';
 
 import { getPaypalSDKMock } from '../mocks/paypal.mock';
 
-import BraintreePaypalCustomerInitializeOptions from './braintree-paypal-customer-initialize-options';
+import BraintreePaypalCustomerInitializeOptions, {
+    WithBraintreePaypalCustomerInitializeOptions,
+} from './braintree-paypal-customer-initialize-options';
 import BraintreePaypalCustomerStrategy from './braintree-paypal-customer-strategy';
 
 describe('BraintreePaypalCustomerStrategy', () => {
@@ -50,10 +55,12 @@ describe('BraintreePaypalCustomerStrategy', () => {
 
     const braintreePaypalOptions: BraintreePaypalCustomerInitializeOptions = {
         container: defaultButtonContainerId,
+        onClick: jest.fn(),
         onError: jest.fn(),
     };
 
-    const initializationOptions: CustomerInitializeOptions = {
+    const initializationOptions: CustomerInitializeOptions &
+        WithBraintreePaypalCustomerInitializeOptions = {
         methodId: 'braintreepaypal',
         braintreepaypal: braintreePaypalOptions,
     };
@@ -133,6 +140,12 @@ describe('BraintreePaypalCustomerStrategy', () => {
             eventEmitter.on('approve', () => {
                 if (typeof options.onApprove === 'function') {
                     options.onApprove({ payerId: 'PAYER_ID' }).catch(() => {});
+                }
+            });
+
+            eventEmitter.on('click', () => {
+                if (typeof options.onClick === 'function') {
+                    options.onClick();
                 }
             });
 
@@ -284,6 +297,7 @@ describe('BraintreePaypalCustomerStrategy', () => {
                 },
                 createOrder: expect.any(Function),
                 onApprove: expect.any(Function),
+                onClick: expect.any(Function),
             });
             expect(renderMock).not.toHaveBeenCalled();
         });
@@ -308,6 +322,7 @@ describe('BraintreePaypalCustomerStrategy', () => {
                 },
                 createOrder: expect.any(Function),
                 onApprove: expect.any(Function),
+                onClick: expect.any(Function),
             });
         });
 
@@ -329,6 +344,7 @@ describe('BraintreePaypalCustomerStrategy', () => {
                 },
                 createOrder: expect.any(Function),
                 onApprove: expect.any(Function),
+                onClick: expect.any(Function),
             });
         });
 
@@ -343,6 +359,18 @@ describe('BraintreePaypalCustomerStrategy', () => {
         });
 
         it('loads checkout details when customer is ready to pay', async () => {
+            jest.spyOn(
+                paymentIntegrationService.getState(),
+                'getPaymentMethodOrThrow',
+            ).mockImplementation(() => {
+                throw new Error();
+            });
+
+            jest.spyOn(paymentIntegrationService, 'loadPaymentMethod').mockResolvedValueOnce({
+                getPaymentMethodOrThrow: jest.fn().mockReturnValue(paymentMethodMock),
+                getStoreConfigOrThrow: jest.fn().mockReturnValue(getConfig().storeConfig),
+                getCartOrThrow: jest.fn().mockReturnValue({ currency: { code: 'AUD' } }),
+            });
             await strategy.initialize(initializationOptions);
 
             eventEmitter.emit('createOrder');