+++ packages/braintree-utils/src/index.ts
@@ -20,6 +20,6 @@ export {
     getBraintreeAddress,
     getBraintreePaypal,
 } from './mocks/braintree.mock';
+export { getValidBraintreeConnectStyles, isBraintreeConnectWindow } from './utils';
 export { getPaypalMock } from './mocks/paypal.mock';
-export { getValidBraintreeConnectStyles } from './utils';
 export { BRAINTREE_SDK_STABLE_VERSION, BRAINTREE_SDK_ALPHA_VERSION } from './sdk-verison';