+++ packages/core/src/analytics/index.ts
@@ -1,2 +1,7 @@
 export { default as createStepTracker } from './create-step-tracker';
 export { default as StepTracker } from './step-tracker';
+
+export {
+    createBraintreeConnectTracker,
+    BraintreeConnectTrackerService,
+} from './braintree-connect-tracker';