+++ packages/core/src/customer/strategies/amazon-pay-v2/amazon-pay-v2-customer-strategy.ts
@@ -1,3 +1,5 @@
+import { PaymentMethod } from '@bigcommerce/checkout-sdk/payment-integration-api';
+
 import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
 import { InvalidArgumentError, NotImplementedError } from '../../../common/error/errors';
 import { PaymentMethodActionCreator } from '../../../payment';
@@ -30,19 +32,32 @@ export default class AmazonPayV2CustomerStrategy implements CustomerStrategy {
             );
         }
 
-        const {
-            paymentMethods: { getPaymentMethodOrThrow },
-        } = await this._store.dispatch(
-            this._paymentMethodActionCreator.loadPaymentMethod(methodId),
-        );
+        if (amazonpay.onClick && typeof amazonpay.onClick !== 'function') {
+            throw new InvalidArgumentError(
+                'Unable to proceed because "options.amazonpay.onClick" argument is not a function.',
+            );
+        }
 
-        await this._amazonPayV2PaymentProcessor.initialize(getPaymentMethodOrThrow(methodId));
+        let state = this._store.getState();
+        let paymentMethod: PaymentMethod<any>;
+
+        try {
+            paymentMethod = state.paymentMethods.getPaymentMethodOrThrow(methodId);
+        } catch (_e) {
+            state = await this._store.dispatch(
+                this._paymentMethodActionCreator.loadPaymentMethod(methodId),
+            );
+            paymentMethod = state.paymentMethods.getPaymentMethodOrThrow(methodId);
+        }
+
+        await this._amazonPayV2PaymentProcessor.initialize(paymentMethod);
 
         this._amazonPayV2PaymentProcessor.renderAmazonPayButton({
             checkoutState: this._store.getState(),
             containerId: amazonpay.container,
             methodId,
             placement: AmazonPayV2Placement.Checkout,
+            ...(amazonpay.onClick && { onClick: amazonpay.onClick }),
         });
 
         return this._store.getState();