+++ packages/core/src/customer/strategies/braintree/braintree-paypal-credit-customer-strategy.ts
@@ -1,6 +1,10 @@
+import { noop } from 'lodash';
 import { FormPoster } from '@bigcommerce/form-poster';
 
-import { DefaultCheckoutButtonHeight } from '@bigcommerce/checkout-sdk/payment-integration-api';
+import {
+    DefaultCheckoutButtonHeight,
+    PaymentMethod,
+} from '@bigcommerce/checkout-sdk/payment-integration-api';
 
 import { CheckoutActionCreator, CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
 import mapToLegacyBillingAddress from '../../../checkout-buttons/strategies/braintree/map-to-legacy-billing-address';
@@ -49,7 +53,6 @@ export default class BraintreePaypalCreditCustomerStrategy implements CustomerSt
 
     async initialize(options: CustomerInitializeOptions): Promise<InternalCheckoutSelectors> {
         const { braintreepaypalcredit, methodId } = options;
-        const { container, buttonHeight } = braintreepaypalcredit || {};
 
         if (!methodId) {
             throw new InvalidArgumentError(
@@ -63,17 +66,25 @@ export default class BraintreePaypalCreditCustomerStrategy implements CustomerSt
             );
         }
 
-        if (!container) {
+        if (!braintreepaypalcredit.container) {
             throw new InvalidArgumentError(
-                `Unable to initialize payment because "braintreepaypalcredit.container" argument is not provided.`,
+                `Unable to initialize payment because "options.braintreepaypalcredit.container" argument is not provided.`,
             );
         }
 
-        const state = await this._store.dispatch(
-            this._paymentMethodActionCreator.loadPaymentMethod(methodId),
-        );
+        let state = this._store.getState();
+        let paymentMethod: PaymentMethod<any>;
+
+        try {
+            paymentMethod = state.paymentMethods.getPaymentMethodOrThrow(methodId);
+        } catch (_e) {
+            state = await this._store.dispatch(
+                this._paymentMethodActionCreator.loadPaymentMethod(methodId),
+            );
+            paymentMethod = state.paymentMethods.getPaymentMethodOrThrow(methodId);
+        }
+
         const storeConfig = state.config.getStoreConfigOrThrow();
-        const paymentMethod = state.paymentMethods.getPaymentMethodOrThrow(methodId);
         const { clientToken, initializationData } = paymentMethod;
 
         if (!clientToken || !initializationData) {
@@ -93,7 +104,6 @@ export default class BraintreePaypalCreditCustomerStrategy implements CustomerSt
                 braintreepaypalcredit,
                 methodId,
                 Boolean(paymentMethod.config.testMode),
-                buttonHeight,
             );
         const paypalCheckoutErrorCallback = (error: BraintreeError) =>
             this._handleError(error, braintreepaypalcredit);
@@ -140,9 +150,12 @@ export default class BraintreePaypalCreditCustomerStrategy implements CustomerSt
         braintreepaypalcredit: BraintreePaypalCreditCustomerInitializeOptions,
         methodId: string,
         testMode: boolean,
-        buttonHeight = DefaultCheckoutButtonHeight,
     ): void {
-        const { container } = braintreepaypalcredit;
+        const {
+            container,
+            buttonHeight = DefaultCheckoutButtonHeight,
+            onClick = noop,
+        } = braintreepaypalcredit;
         const { paypal } = this._window;
 
         let hasRenderedSmartButton = false;
@@ -179,6 +192,7 @@ export default class BraintreePaypalCreditCustomerStrategy implements CustomerSt
                                 braintreepaypalcredit,
                                 methodId,
                             ),
+                        onClick,
                     });
 
                     if (paypalButtonRender.isEligible()) {