+++ packages/core/src/payment/strategies/amazon-pay-v2/amazon-pay-v2-payment-processor.ts
@@ -64,11 +64,23 @@ export default class AmazonPayV2PaymentProcessor {
         });
     }
 
-    createButton(containerId: string, options: AmazonPayV2ButtonParameters): void {
+    createButton(
+        containerId: string,
+        options: AmazonPayV2ButtonParameters,
+        onClick?: () => void,
+    ): void {
         this._amazonPayV2Button = this._getAmazonPayV2SDK().Pay.renderButton(
             `#${containerId}`,
             options,
         );
+
+        // onClick callback is a function provided through initializationOptions to the strategy
+        // P.S.: related to customer strategy
+        if (onClick && typeof onClick === 'function') {
+            this._amazonPayV2Button.onClick(() => {
+                onClick();
+            });
+        }
     }
 
     prepareCheckout(createCheckoutSessionConfig: Required<AmazonPayV2CheckoutSessionConfig>) {
@@ -120,6 +132,7 @@ export default class AmazonPayV2PaymentProcessor {
         methodId,
         options,
         placement,
+        onClick,
     }: AmazonPayV2ButtonRenderingOptions): HTMLDivElement {
         const container = document.querySelector<HTMLElement>(`#${containerId}`);
 
@@ -145,7 +158,7 @@ export default class AmazonPayV2PaymentProcessor {
                 buttonColor,
             );
 
-        this.createButton(parentContainerId, amazonPayV2ButtonOptions);
+        this.createButton(parentContainerId, amazonPayV2ButtonOptions, onClick);
 
         return this._getButtonParentContainer();
     }