+++ packages/paypal-commerce-integration/src/paypal-commerce-credit/paypal-commerce-credit-customer-initialize-options.ts
@@ -16,6 +16,11 @@ export default interface PayPalCommerceCreditCustomerInitializeOptions {
      * A callback that gets called when payment complete on paypal side.
      */
     onComplete?(): void;
+
+    /**
+     * A callback that gets called when paypal button clicked.
+     */
+    onClick?(): void;
 }
 
 export interface WithPayPalCommerceCreditCustomerInitializeOptions {